+++ workflows_definition/src/sdk/closing_reason.py
@@ -26,7 +26,7 @@ class ClosingReason:
         
         url = utils.generate_url(operations.ChangeReasonStatusRequest, base_url, '/v1/workflows/closing-reasons/{reasonId}', request)
         headers = {}
-        req_content_type, data, form = utils.serialize_request_body(request, "change_reason_status_req", False, True, 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, operations.ChangeReasonStatusRequest, "change_reason_status_req", False, True, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         headers['Accept'] = 'application/json'
@@ -66,7 +66,7 @@ class ClosingReason:
         
         url = base_url + '/v1/workflows/closing-reasons'
         headers = {}
-        req_content_type, data, form = utils.serialize_request_body(request, "request", False, False, 'json')
+        req_content_type, data, form = utils.serialize_request_body(request, shared.ClosingReason, "request", False, False, 'json')
         if req_content_type not in ('multipart/form-data', 'multipart/mixed'):
             headers['content-type'] = req_content_type
         if data is None and form is None: