+++ workflows_definition/src/sdk/sdkconfiguration.py
@@ -21,9 +21,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '2.1.4'
-    gen_version: str = '2.210.6'
-    user_agent: str = 'speakeasy-sdk/python 2.1.4 2.210.6 1.0.0 openapi'
+    sdk_version: str = '2.2.0'
+    gen_version: str = '2.213.3'
+    user_agent: str = 'speakeasy-sdk/python 2.2.0 2.213.3 1.0.0 openapi'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: